fix(core): move `available_cpu_cores` to `runtime.rs`
